AVX optimized norm function implementations migrated to separate file
